Backend/User/requirements.txt
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
flask-mongoengine==1.0.0
Flask-WTF==1.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.1
pycparser==2.21
PyJWT==2.4.0
pymongo==4.1.1
requests==2.28.0
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
gunicorn==20.0.4
Flask-Cors==3.0.10
marshmallow==3.17.0
